requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
pymongo 
python-dotenv